mar2026-API/requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
asttokens==3.0.1
attrs==25.4.0
Beaker==1.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
bokeh==3.8.2
boto3==1.35.9
botocore==1.35.9
Brlapi==0.8.7
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cockpit @ file:///builddir/build/BUILD/cockpit-357-build/cockpit-357/tmp/wheel/cockpit-357-py3-none-any.whl#sha256=474f6d2f07f5e1fc6d7a1714574c3620ecb9769fd645f220ab4b2314f02dc195
comm==0.2.3
contourpy==1.3.3
crypt_r==3.13.1
cryptography==45.0.4
cupshelpers==1.0
dasbus==1.7
dask==2026.1.2
dbus-python==1.4.0
dbus_next==0.2.3
debugpy==1.8.20
decorator==5.2.1
distributed==2026.1.2
distro==1.9.0
dnf==4.24.0
executing==2.2.1
fastapi==0.135.1
fedora-third-party==0.10
file-magic==0.4.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.10
langchain-core==1.2.16
langchain-openai==1.1.10
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
locket==1.0.0
louis==3.33.0
lxml==6.0.1
lz4==4.4.5
Mako==1.2.3
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
msgpack==1.1.2
narwhals==2.17.0
nest-asyncio==1.6.0
nftables==0.1
-e git+https://github.com/andre-servicos-gmn/NV09224802.git@66cfa9a2ecf8ecddbd2c4b78264e8d9ce3f921f4#egg=nouvaris_agents_v2&subdirectory=backend
numpy==2.4.2
olefile==0.47
openai==2.24.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
parso==0.8.6
partd==1.4.2
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
pluggy==1.6.0
ply==3.11
productmd==1.50
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pyOpenSSL==25.1.0
pyparted==3.13.0
PySocks==1.7.1
pytest==9.0.2
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2025.2
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.12.2
referencing==0.37.0
regex==2026.2.28
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
rpm==6.0.1
ruff==0.15.4
s3transfer==0.10.4
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.11.0
soupsieve==2.8
stack-data==0.6.3
starlette==0.52.1
streamlit==1.54.0
systemd-python==235
tblib==3.2.2
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.6.0
websockets==16.0
xkbregistry==0.3
xxhash==3.6.0
xyzservices==2025.11.0
zict==3.0.0
zstandard==0.25.0
